update install.sh to make installation process more easier & add redis service

+++ install.sh
@@ -1,8 +1,24 @@
-echo "Coping docker-compose.yml.example to docker-compose.yml"
-cp docker-compose.yml.example docker-compose.yml
+#!/bin/bash
 
-echo "Coping .env.example to .env"
-cp .env.example .env
+#if already have docker-compose.yml and .env, then skip copy from example 
+if [ -f "docker-compose.yml" ]; then
+    echo "docker-compose.yml already exists"
+else
+    echo "Coping docker-compose.yml.example to docker-compose.yml"
+    cp docker-compose.yml.example docker-compose.yml
+fi
+
+if [ -f ".env" ]; then
+    echo ".env already exists"
+else
+    echo "Coping .env.example to .env"
+    cp .env.example .env
+fi
+
+read -p "Do you want to review the env file? y/n: " CHOICE
+if [[ $CHOICE == "y" ]]; then
+	nano .env
+fi
 
 #laravel app
 echo "Cloning latest version of laravel"
